[CLOUD-2894] Port simple WF 14 changes to standalone-openshift.xml (#29)
Note this does not include the WFLY-10541 change discussed at https://lists.jboss.org/pipermail/wildfly-dev/2018-July/006601.html. That change was to add an expression to the standard WildFly configs to allow the id to be set easily via the command line. The standalone-openshift.xml already supports this with a value of ${jboss.node.name} which will be unique per pod. Changing that to the WFLY-10541 default of ${jboss.tx.node.id:1} will break things as jboss.tx.node.id will not be set and '1' is not an appropriate value.

It also does not include the WFLY-10614 change, which, if confirmed appropriate, will be brought in separately.

Signed-off-by: Brian Stansberry <brian.stansberry@redhat.com>